汎用的なファイル処理を関数に切り出し
